edit: variable management
