+++ tests/env_vars/post_create_user.http
@@ -16,10 +16,10 @@ Authorization: Bearer {{TOKEN}}
     "id" : "{{$uuid}}"
 }
 
-----
+###
 
 POST {{URL}}
 Content-Type: application/json
 Authorization: Bearer {{TOKEN}}
 
-< {{PAYLOAD]}}
+< {{PAYLOAD}}
