+++ tests/script_vars/.env
@@ -1,2 +1,1 @@
-postId=3
-userId=1
+postTitle="temporibus sit alias delectus eligendi possimus magni"
